+++ src/main/java/Pimod/card/testCard/extendstest2.java
@@ -1,4 +1,4 @@
-package Pimod.card;
+package Pimod.card.testCard;
 
 import Pimod.patches.AbstractCardEnum;
 import com.megacrit.cardcrawl.actions.AbstractGameAction;
@@ -13,8 +13,8 @@ import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import basemod.abstracts.CustomCard;
 
-public class extendstest122 extends CustomCard {
-    public static final String ID = "extendstest122";
+public class extendstest2 extends CustomCard {
+    public static final String ID = "extendstest2";
     private static CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
     // Get object containing the strings that are displayed in the game.
     public static final String NAME = cardStrings.NAME;
@@ -26,10 +26,10 @@ public class extendstest122 extends CustomCard {
     private static final int VULNERABLE_AMT = 1;
     private static final int UPGRADE_PLUS_VULNERABLE = 1;
 
-    public extendstest122() {
+    public extendstest2() {
         super(ID, NAME, IMG_PATH, COST, DESCRIPTION,
                 CardType.POWER, AbstractCardEnum.PI_DERIVATIONS,
-                CardRarity.RARE, AbstractCard.CardTarget.ENEMY);
+                CardRarity.UNCOMMON, AbstractCard.CardTarget.ENEMY);
         this.magicNumber = this.baseMagicNumber = VULNERABLE_AMT;
         this.damage=this.baseDamage = ATTACK_DMG;
 
@@ -45,7 +45,7 @@ public class extendstest122 extends CustomCard {
 
     @Override
     public AbstractCard makeCopy() {
-        return new extendstest122();
+        return new extendstest2();
     }
 
     @Override
