+++ src/main/java/Pimod/tutorial/Tutorial.java
@@ -1,36 +1,29 @@
 package Pimod.tutorial;
 import Pimod.card.*;
+import Pimod.card.PiBaseCard.Defend_PI;
+import Pimod.card.PiBaseCard.Strike_PI;
+import Pimod.card.testCard.*;
 import Pimod.cardActions.addPiCards;
 import Pimod.characters.A_PI;
-import Pimod.campfire.*;
 import Pimod.patches.AbstractCardEnum;
 import Pimod.patches.PIClassEnum;
 import Pimod.relic.goldenApple;
 import basemod.BaseMod;
-import basemod.abstracts.CustomRelic;
-import basemod.helpers.RelicType;
 import basemod.interfaces.*;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.utils.compression.lzma.Base;
 import com.evacipated.cardcrawl.modthespire.lib.SpireInitializer;
 import com.google.gson.Gson;
 import com.megacrit.cardcrawl.cards.AbstractCard;
-import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.helpers.CardHelper;
 import com.megacrit.cardcrawl.localization.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import com.megacrit.cardcrawl.relics.AbstractRelic;
-import com.megacrit.cardcrawl.ui.campfire.AbstractCampfireOption;
-import com.megacrit.cardcrawl.ui.campfire.DigOption;
-import org.apache.commons.codec.Charsets;
-import static Pimod.cardActions.getRandomPiCards.PiCardPool;
+
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Iterator;
-import Pimod.powers.withered;
 
 
 @SpireInitializer
